refactor: 랜덤 파워 생성 리팩터링
